Move chunk.py to top level
